+++ src/main/java/com/example/springboot3/Dao/ReservationDao.java
@@ -10,12 +10,12 @@ import static com.yeji.jooq.generated.tables.ProductPrice.PRODUCT_PRICE;
 import static com.yeji.jooq.generated.tables.Promotion.PROMOTION;
 import static com.yeji.jooq.generated.tables.ReservationUserComment.RESERVATION_USER_COMMENT;
 
-import com.example.springboot3.Dto.categoryApi.CategoryItemDTO;
-import com.example.springboot3.Dto.detailApi.DisplayInfoImageDTO;
-import com.example.springboot3.Dto.detailApi.ProductImageDTO;
-import com.example.springboot3.Dto.detailApi.ProductPriceDTO;
-import com.example.springboot3.Dto.displayInfoApi.DisplayInfoItemDTO;
-import com.example.springboot3.Dto.promotionApi.PromotionItemDTO;
+import com.example.springboot3.Dto.CategoryItemDTO;
+import com.example.springboot3.Dto.DisplayInfoImageDTO;
+import com.example.springboot3.Dto.DisplayInfoItemDTO;
+import com.example.springboot3.Dto.ProductImageDTO;
+import com.example.springboot3.Dto.ProductPriceDTO;
+import com.example.springboot3.Dto.PromotionItemDTO;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.jooq.DSLContext;
