+++ src/main/java/com/example/springboot3/Dto/ProductImageDTO.java
@@ -1,4 +1,4 @@
-package com.example.springboot3.Dto.detailApi;
+package com.example.springboot3.Dto;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import java.time.LocalDateTime;
@@ -7,18 +7,21 @@ import lombok.Getter;
 
 @Getter
 @AllArgsConstructor
-public class ProductPriceDTO {
+public class ProductImageDTO {
 
-    private int id;
     private int productId;
-    private String priceTypeName;
-    private int price;
-    private int discountRate;
+    private int productImageId;
+    private String type;
+    private int fileInfoId;
+    private String fileName;
+    private String saveFileName;
+    private String contentType;
+    private int deleteFlag;
 
     @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss.S")
     private LocalDateTime createDate;
 
     @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss.S")
     private LocalDateTime modifyDate;
-    
+
 }
