+++ src/main/java/com/example/springboot3/Dto/PromotionItemDTO.java
@@ -1,4 +1,4 @@
-package com.example.springboot3.Dto.promotionApi;
+package com.example.springboot3.Dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
