+++ src/main/java/com/example/springboot3/Dto/response/DisplayInfoResponseDTO.java
@@ -1,5 +1,6 @@
-package com.example.springboot3.Dto.displayInfoApi;
+package com.example.springboot3.Dto.response;
 
+import com.example.springboot3.Dto.DisplayInfoItemDTO;
 import java.util.List;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -11,6 +12,5 @@ public class DisplayInfoResponseDTO {
     private int totalCount;
     private int productCount;
     private List<DisplayInfoItemDTO> displayInfoItemList;
-
 
 }
